Split to new module crypto
