+++ build-logic/src/main/kotlin/BuildModules.kt
@@ -1,0 +1,6 @@
+/**
+ * Configuration of build modules
+ */
+object BuildModules {
+    const val CRYPTO = ":crypto"
+}
